# src/main/java/org/muzika/playlistmanager/dto/PlaylistResponse.java
package org.muzika.playlistmanager.dto;

import lombok.Data;

import java.time.LocalDateTime;
import java.util.UUID;

@Data
public class PlaylistResponse {
    private UUID id;
    private UUID userId;
    private String name;
    private String description;
    private LocalDateTime createdAt;
    private LocalDateTime updatedAt;
}